# apps/web/src/shared/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet Maintenance Decision API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/plans/{plan_id}/comparison": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Compare Saved Plan",
        "operationId": "compare_saved_plan_api_plans__plan_id__comparison_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_plans_get",
        "parameters": [
          {
            "name": "scope_component_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Component Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponse"
                  },
                  "title": "Response List Plans Api Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/proposals": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Proposal",
        "operationId": "create_proposal_api_plans_proposals_post",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/approve": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Approve",
        "operationId": "approve_api_plans__plan_id__approve_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/revisions": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Revise",
        "operationId": "revise_api_plans__plan_id__revisions_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/rejection": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Reject",
        "operationId": "reject_api_plans__plan_id__rejection_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}/commitment": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Commitment",
        "operationId": "commitment_api_plans__plan_id__commitment_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanCommitmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "attempt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "state",
          "attempt",
          "result"
        ],
        "title": "JobResponse"
      },
...
      "PlanAssignmentResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "start": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End"
          },
          "crew_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crew Id"
          },
          "bay_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bay Id"
          },
          "crew_unit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Crew Unit",
            "default": 0
          },
          "bay_unit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bay Unit",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "start",
          "end"
        ],
        "title": "PlanAssignmentResponse"
      },
      "PlanCommitmentResponse": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "reservations": {
            "items": {
              "$ref": "#/components/schemas/ReservationHistoryResponse"
            },
            "type": "array",
            "title": "Reservations"
          },
          "work": {
            "items": {
              "$ref": "#/components/schemas/WorkHistoryResponse"
            },
            "type": "array",
            "title": "Work"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "reservations",
          "work"
        ],
        "title": "PlanCommitmentResponse"
      },
      "PlanResponse": {
        "properties": {
          "input_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Snapshot"
          },
          "objective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "best_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Bound"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "solver_status": {
            "type": "string",
            "title": "Solver Status"
          },
          "input_version": {
            "type": "string",
            "title": "Input Version"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/PlanAssignmentResponse"
            },
            "type": "array",
            "title": "Assignments"
          },
          "diagnostics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "solver_status",
          "input_version",
          "assignments",
          "diagnostics",
          "created_at",
          "approved_at",
          "approved_by"
        ],
        "title": "PlanResponse"
      },
...
      "ReservationHistoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "part_id": {
            "type": "string",
            "title": "Part Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "part_id",
          "quantity",
          "status"
        ],
        "title": "ReservationHistoryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkHistoryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "consumed_quantity": {
            "type": "integer",
            "title": "Consumed Quantity"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "status",
          "version",
          "consumed_quantity",
          "notes",
          "started_at",
          "completed_at"
        ],
        "title": "WorkHistoryResponse"
      },
...
    }
  }
}